provider: remove the destination parameter from Transcode method
This is usually defined in the provider. The provider is configured to
set the transcoded file to somewhere. For Encoding.com, it will be a
configuration. For Elastic Transcoder, the destination bucket will be
predefined in the pipeline.

For the moment, we care only about S3. We should be able to handle other
destinations through Encoding.com, but not now :-)