db: add Preset persistence
As part of #7, we want to be able to persist presets on the database.

I've also added some "magic" on the db package for automatically saving
and loading types using Redis hashes. It's limited to strings at the
moment, meaning that we can persist structs and maps, as long as their
keys and values are strings. Here is an example of valid value:

	type Person struct {
		ID      string  `redis-hash:"-"`
		Name    string  `redis-hash:"name"`
		Address Address `redis-hash:",expand"`
	}

	type Address struct {
		Data map[string]string `redis-hash:",expand"`
	}

Unexported fields are ignored, even if they declare the redis-hash tag.
There are still some cases to handle, like absent struct tag, and
conflicting fields (maybe we could have an automatic prefix on
expansion?).