db/redis: implement the ListPresets method
This is bizarre because we're using Redis, so we need an atomic
operation to create the preset (the preset is an entry in a Redis set
and also a hash).

On delete, we're not caring about the atomicity of the operation, we
just want to ensure that the hash is deleted, if the id is living in the
set, that's fine, because List will ignore not found error.

It's important to keep in mind that adding a new preset is a slow
operation, as well as listing them, but we expect to be using the
presets more than listing or creating new presets, so that should be
fine.