service: add swagger notations to all preset endpoints
I changed a lot the structure of the code. There's a new type for
representing our handlers now. This type is a function that takes a
request and returns a gizmoResponse.

gizmoResponse is an interface with a single method: Result(). Result
returns the three parameters required by a Gizmo JSONEndpoint. Then we
can have structs implementing the gizmoResponse interface, and annotate
these structs with swagger:response.

Things are looking better now, but I should definitely get some code
review here.