bin/build: change build options
Do not compile dependencies when downloading then (the flag -d means
"download only"), and do not download test dependencies. We're not going
to run tests for now. I understand the impact of that: since we're
blindly running `go get`, we're always downloading the latest version of
our dependencies, so running tests is a safe path. I still think that an
even safer path is to have dependencies vendored, so might be something
that we tackle in the future.

I've also included the swagger.json file in the tarball.