bin/build: speed-up local build time
In case we have GOPATH defined, we probably have the dependencies, than
we don't need to download them again. It still works as expected in a
box with no GOPATH defined.

It's way faster to avoid downloading the dependencies. This is because
`go get` is dumb when cloning Git repositories. This should probably
change when golang/go#13078 is fixed (Go 1.7, hopefully).