swagger: create package with shared types and methods
I wanted to share more than that (specially baseResp), but go-swagger
doesn't handle anonymous structs as I would like it too. I tried
embedding an interface, but that didn't work either, so I just gave up.
Almost everything can be shared with other libraries, but we still need
to duplicate some code (copy-n-paste baseResp in all projects and
implement Result in error types, for example).

If go-swagger starts handling embedded structs or interfaces in the
future, we can improve this code further, by having an interface with
the method Result() and embedding it around. Then we would be able to
move baseResp to the swagger package (and call it "Response").

That would allow us to remove the double Result definition in error
responses, and also all the replication os baseResp. I opened
go-swagger/go-swagger#493 to address interface embedding.