Make the output filename a parameter in the API
This way, the filename is a parameter to the Transcode method. The
transcode method now takes a list of "outputs" and each output is a
struct with the preset and the desired file name. This change also
includes a new attribute in the StreamingParams struct
(PlaylistFileName). This should refer to the master file of the
streaming job.

This change also expands the input for the newTranscodeJob to include
the output filename and the playlist file name. Both parameters are
optional and have the following default values:

- output.fileName: {source file}_{preset name}.{preset extension}
- streamingParams.playlistFileName: hls/index.m3u8

With this change, we no longer need the label in the preset, so I
removed it from the db layer.

In Elemental Conductor, it's important to notice that for the HLS output
group, we ignore the filename in each level preset. The code only uses
streamingParams.playlistFileName and then attaches the _000000000x
suffix. It has the side effect that we only support 9999999999 levels in
a master playlist :-D