refactor: change check for cfg creds
We don't want this code to do anything inexpected.

By checking these values separately, if a config is incomplete and a key was provided but no secret (or the other way around) the static creds logic is not run and when run, nothing will error and default creds may be used unexpectedly. Let's make sure that if anything was provided for key or secret, we run the static credentials logic.